+++ CHANGELOG.md
@@ -8,6 +8,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Support for streams and buffers in `pyflp.parse` and `pyflp.save` [#155].
+- `FLVersion.from_str` to parse an FL Studio version string.
+- Basic logging when an event fails to parse.
+
+[#155]: https://github.com/demberto/PyFLP/issues/155
+
 ## [2.2.1] - 2023-06-05
 
 ### Fixed